feat/refactoring
