+++ lrml-visual-editor/src/blockly/blocks.js
@@ -115,6 +115,10 @@ Blockly.defineBlocksWithJsonArray([
           [
             "by",
             "by"
+          ],
+          [
+            "exist",
+            "exist"
           ]
         ]
       }
